+++ Nickel/Framework/Content/ContentManager.cs
@@ -51,7 +51,7 @@ internal sealed class ContentManager
 		var statuses = new StatusManager(currentModLoadPhaseProvider, vanillaModManifest);
 		var cards = new CardManager(currentModLoadPhaseProvider, loggerProvider, vanillaModManifest);
 		var artifacts = new ArtifactManager(currentModLoadPhaseProvider, loggerProvider, vanillaModManifest);
-		var characters = new CharacterManager(currentModLoadPhaseProvider, loggerProvider, sprites, decks, statuses);
+		var characters = new CharacterManager(currentModLoadPhaseProvider, loggerProvider, sprites, decks, statuses, vanillaModManifest);
 		var parts = new PartManager(currentModLoadPhaseProvider);
 		var ships = new ShipManager(currentModLoadPhaseProvider);
 		var cardTraits = new CardTraitManager(vanillaModManifest, modManagerModManifest, modDataManager);